requirements.txt
astor==0.8.1
attrs==19.3.0
Automat==0.8.0
Babel==2.9.1
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.62
blinker==1.4
cachetools==4.2.4
certifi==2019.11.28
cfgv==3.3.1
chardet==3.0.4
click==8.0.1
#cloud-init==20.1
colorama==0.4.3
#command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
distlib==0.3.3
distro==1.4.0
#distro-info===0.23ubuntu1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.3.0
flake8==3.9.2
Flask==1.1.2
Flask-Babel==2.0.0
future==0.18.2
gast==0.4.0
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
identify==2.3.0
idna==2.8
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.5.0
incremental==16.10.1
iopath==0.1.9
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.15.2
Jinja2==3.0.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.2
#language-selector==0.1
#launchpadlib==1.10.13
#layoutparser==0.0.0
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lmdb==1.2.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.6.3
nodeenv==1.6.0
numpy==1.21.2
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.5.3.56
openpyxl==3.0.9
paddleocr==2.3.0.1
paddlepaddle==2.1.3
pandas==1.3.3
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
portalocker==2.3.2
pre-commit==2.15.0
premailer==3.10.0
prompt-toolkit==2.0.10
protobuf==3.18.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyclipper==1.3.0
pycodestyle==2.7.0
pycryptodome==3.10.4
pyflakes==2.3.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
#python-apt==2.0.0
python-dateutil==2.8.2
#python-debian===0.1.36ubuntu1
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
#scikit-image==0.17.2
scipy==1.7.1
SecretStorage==2.3.1
service-identity==18.1.0
Shapely==1.7.1
shellcheck-py==0.7.2.1
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
systemd-python==234
tifffile==2021.8.30
toml==0.10.2
tqdm==4.62.3
traitlets==4.3.3
Twisted==18.9.0
#ubuntu-advantage-tools==20.3
#ufw==0.36
#unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.8.1
visualdl==2.2.1
wadllib==1.3.3
wcwidth==0.1.8
Werkzeug==2.0.2
zipp==1.0.0
#zope.interface==4.7.1
